fix: instantiate compute_{min,max}_positions_int8 with i8, not i64
generic_compute!(compute_min_positions_int8, ...) and its max
counterpart were instantiated with i64 despite the _int8 name, so the
generated #[pyfunction] expected a PyReadonlyArray1<i64> argument
instead of PyReadonlyArray1<i8>, breaking dispatch for actual int8
numpy arrays. Swept the rest of the aggs tree for the same typo;
every other _int8 instantiation already uses i8 correctly.

Fixes #30

+++ src/aggs/max/max_positions.rs
@@ -53,7 +53,7 @@ macro_rules! generic_compute {
 generic_compute!(compute_max_positions_int64, i64);
 generic_compute!(compute_max_positions_int32, i32);
 generic_compute!(compute_max_positions_int16, i16);
-generic_compute!(compute_max_positions_int8, i64);
+generic_compute!(compute_max_positions_int8, i8);
 generic_compute!(compute_max_positions_uint64, u64);
 generic_compute!(compute_max_positions_uint32, u32);
 generic_compute!(compute_max_positions_uint16, u16);
